+++ data/src/main/java/com/nicolas/data/preferences/Preferences.kt
@@ -10,7 +10,7 @@ interface Preferences {
     fun saveGender(gender: Gender)
     fun saveAge(age: Int)
     fun saveWeight(weight: Float)
-    fun saveHeight(height: Float)
+    fun saveHeight(height: Int)
     fun saveActivityLevel(level: ActivityLevel)
     fun saveGoalType(type: GoalType)
     fun saveCarbsRatio(ratio: Float)
